Add conditions to like and dislike

+++ post.php
@@ -27,60 +27,59 @@ if (isset($_POST['deleted'])) {
 			echo "1";
 			exit();
 		}
-
 		else {
 			exit();
 		}
 	}
-
 	else {
 		exit();
 	}
 }
 
 if (isset($_POST['liked'])) {
-	$postid = $_POST['postid'];
-	$result = mysqli_query($conn, "SELECT * FROM posts WHERE id=$postid");
-	$row = mysqli_fetch_array($result);
-	$n = $row['likes'];
+	if ($user) {
+		$postid = $_POST['postid'];
+		$result = mysqli_query($conn, "SELECT * FROM posts WHERE id=$postid");
+		$row = mysqli_fetch_array($result);
+		$n = $row['likes'];
 
-	$liked = mysqli_query($conn, "SELECT * FROM likes WHERE username='$user' AND postid=$postid");
-	$likedrows = mysqli_num_rows($liked);
+		$liked = mysqli_query($conn, "SELECT * FROM likes WHERE username='$user' AND postid=$postid");
+		$likedrows = mysqli_num_rows($liked);
 
-    if ($likedrows == 0) {
-        mysqli_query($conn, "INSERT INTO likes (id, username, postid) VALUES (NULL, '$user', $postid)");
-	    mysqli_query($conn, "UPDATE posts SET likes=$n+1 WHERE id=$postid");
-		echo $n+1;
-		exit();
-    }
-
+		if ($likedrows == 0) {
+			mysqli_query($conn, "INSERT INTO likes (id, username, postid) VALUES (NULL, '$user', $postid)");
+			mysqli_query($conn, "UPDATE posts SET likes=$n+1 WHERE id=$postid");
+			echo $n+1;
+			exit();
+		}
+		else {
+			exit();
+		}
+	}
 	else {
-		echo $n;
-		echo "You have already liked this post.";
 		exit();
 	}
 }
 
 if (isset($_POST['unliked'])) {
-	$postid = $_POST['postid'];
-	$result = mysqli_query($conn, "SELECT * FROM posts WHERE id=$postid");
-	$row = mysqli_fetch_array($result);
-	$n = $row['likes'];
+	if ($user) {
+		$postid = $_POST['postid'];
+		$result = mysqli_query($conn, "SELECT * FROM posts WHERE id=$postid");
+		$row = mysqli_fetch_array($result);
+		$n = $row['likes'];
 
-	$liked = mysqli_query($conn, "SELECT * FROM likes WHERE username='$user' AND postid=$postid");
-	$likedrows = mysqli_num_rows($liked);
+		$liked = mysqli_query($conn, "SELECT * FROM likes WHERE username='$user' AND postid=$postid");
+		$likedrows = mysqli_num_rows($liked);
 
-    if ($likedrows == 0) {
-        echo $n;
-		echo "You have not liked this post.";
-		exit();
-    }
-
-	else {
-		mysqli_query($conn, "DELETE FROM likes WHERE postid=$postid AND username='$user'");
-		mysqli_query($conn, "UPDATE posts SET likes=$n-1 WHERE id=$postid");
-		echo $n-1;
-		exit();
-		
+		if ($likedrows == 0) {
+			exit();
+		}
+		else {
+			mysqli_query($conn, "DELETE FROM likes WHERE postid=$postid AND username='$user'");
+			mysqli_query($conn, "UPDATE posts SET likes=$n-1 WHERE id=$postid");
+			echo $n-1;
+			exit();
+			
+		}
 	}
 }
